Initial version of buildscript
Made the initial version of the buildscript that does the following:
- possible clean up Release folder
- compile less
- compile js
- copy required files to Release folder
- substritute:
* css references (commented for now, since already correct)
* js references
* PHP settings
* cron settings

Also in this commit is the move of specific build scripts to build fodler
and added the release folder to gitignore.

In the mean time there are also some small fixes w.r.t. settings in
various files.

See #40