Added logging and fixed undefined error.
During the submission of the form on all replies there was a header
statement to indicate a bad request or succesfullr eply. This hampered JS
to parse the message and we got an undefined message.

Added logging for the cases where something goes worng and if there is a
DB error even to send the webmasters an email.

See #49